Change manchester parser to parse to owl ontology (#71)
* Updated visitors in parser

* Finished manchester parser to owlontology

* Transformed manchester based integraiton tests to use pareser + translator

* Fixed Manchester tests

* Formatting

* Fixed prefixes and ontology versions in manchester parser

* Created new manchester 2 alc test project

* Almost got manchester to owl tests working

* Working through compilation issues in manchester tests

* Manchester parser tests compiles

* Changed to biginteger for xsd:integer

* Fixed xsd namespaces error

* Manchester 2 OWL 2 ALC tests work

* Upgraded action to use v9 dotnet runtime

* Added usage examples to README.md

* removed commented out codes

+++ src/Turtle.Parser/ResourceVisitor.cs
@@ -8,6 +8,7 @@
 
 
 using System.Globalization;
+using System.Numerics;
 using DagSemTools.Rdf;
 using DagSemTools.Ingress;
 
@@ -183,7 +184,7 @@ internal class ResourceVisitor : TurtleDocBaseVisitor<uint>
             Namespaces.XsdString => RdfLiteral.NewLiteralString(literalString),
             Namespaces.XsdDouble => RdfLiteral.NewDoubleLiteral(double.Parse(literalString, CultureInfo.InvariantCulture)),
             Namespaces.XsdDecimal => RdfLiteral.NewDecimalLiteral(decimal.Parse(literalString, CultureInfo.InvariantCulture)),
-            Namespaces.XsdInteger => RdfLiteral.NewIntegerLiteral(int.Parse(literalString)),
+            Namespaces.XsdInteger => RdfLiteral.NewIntegerLiteral(BigInteger.Parse(literalString)),
             Namespaces.XsdFloat => RdfLiteral.NewFloatLiteral(float.Parse(literalString, CultureInfo.InvariantCulture)),
             Namespaces.XsdBoolean => RdfLiteral.NewBooleanLiteral(bool.Parse(literalString)),
             Namespaces.XsdDateTime => RdfLiteral.NewDateTimeLiteral(DateTime.Parse(literalString)),
